♻️ 모달 state, hooks 리팩터링
