--- src/hooks/noticeModal.ts
@@ -1,6 +1,0 @@
-import { noticeModalState } from "@state/noticeModal/atoms";
-import { useRecoilState } from "recoil";
-
-export const useNoticeModalState = () => {
-  return useRecoilState(noticeModalState);
-};
