--- src/state/loginModal/atoms.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const loginModalState = atom<boolean>({
-  key: "loginModalState",
-  default: false,
-});
