--- src/state/noticeModal/atoms.ts
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const noticeModalState = atom<boolean>({
-  key: "noticeModalState",
-  default: false,
-});
